% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCID.to.data.frame.R
\name{SCID.to.data.frame}
\alias{SCID.to.data.frame}
\title{Reads all the records in a SCID file (Version 1).}
\usage{
SCID.to.data.frame(filename)
}
\arguments{
\item{filename}{(character) path to a scid file.}
}
\value{
data.frame with columns: DateTime, Open, High, Low, Last, NumberOfTrades, Volume, BidVolume, AskVolume
}
\description{
First reads the header and checks for file format and version.
Then computes the number of records in the file and reads them one by one.
Finally, merges all the records in one data.frame.
See Intraday Data File Format documentation for details about this file format.
}
\author{
Juan Luis Herrera Cortijo \email{juan.luis.herrera.cortijo@gmail.com}
}
\references{
Intraday Data File Format: \url{https://www.sierrachart.com/index.php?page=doc/IntradayDataFileFormat.html}
}

